refactoring the test folder strucutre and naming
