+++ framework/src/processing/cicd-pipeline/index.ts
@@ -2,6 +2,6 @@
 // SPDX-License-Identifier: MIT-0
 
 
-export { SparkEmrCICDPipeline, SparkEmrCICDPipelineProps, SparkImage } from './spark-cicd-pipeline';
+export { SparkEmrCICDPipeline, SparkEmrCICDPipelineProps, SparkImage } from './spark-emr-cicd-pipeline';
 export { ApplicationStackFactory } from '../../utils/application-stack-factory';
 export { CICDStage } from '../../utils/application-stage';
