+++ framework/src/processing/cicd-pipeline/spark-emr-cicd-pipeline.ts
