+++ framework/src/processing/spark-runtime/emr-serverless/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-export { SparkEmrServerlessRuntime } from './spark-runtime-serverless';
-export { SparkEmrServerlessRuntimeProps, Architecture } from './spark-runtime-serverless-props';
+export { SparkEmrServerlessRuntime } from './spark-emr-runtime-serverless';
+export { SparkEmrServerlessRuntimeProps, Architecture } from './spark-emr-runtime-serverless-props';
